adding changes to support custom tmpdir

+++ sregistry/utils/fileio.py
@@ -35,9 +35,9 @@ import io
 import sys
 
 
-############################################################################
-## FOLDER OPERATIONS #######################################################
-############################################################################
+################################################################################
+## FOLDER OPERATIONS ###########################################################
+################################################################################
 
 
 def mkdir_p(path):
@@ -54,24 +54,76 @@ def mkdir_p(path):
             sys.exit(1)
 
 
+def get_tmpfile(requested_tmpdir=None, prefix=""):
+    '''get a temporary file with an optional prefix. By default will be
+       created in /tmp unless SREGISTRY_TMPDIR is set. By default, the file
+       is closed (and just a name returned).
+
+       Parameters
+       ==========
+       requested_tmpdir: an optional requested temporary directory, first
+       priority as is coming from calling function.
+       prefix: Given a need for a sandbox (or similar), prefix the file
+       with this string.
+    '''
+
+    # First priority for the base goes to the user requested.
+    tmpdir = get_tmpdir(requested_tmpdir)
+
+    # If tmpdir is set, add to prefix
+    if tmpdir is not None:
+        prefix = os.path.join(tmpdir, os.path.basename(prefix))
+
+    fd, tmp_file = tempfile.mkstemp(prefix=prefix) 
+    os.close(fd)
+
+    return tmp_file
+
+
+def get_tmpdir(requested_tmpdir=None, prefix="", create=True):
+    '''get a temporary directory for an operation. If SREGISTRY_TMPDIR
+       is set, return that. Otherwise, return the output of tempfile.mkdtemp
+
+       Parameters
+       ==========
+       requested_tmpdir: an optional requested temporary directory, first
+       priority as is coming from calling function.
+       prefix: Given a need for a sandbox (or similar), we will need to 
+       create a subfolder *within* the SREGISTRY_TMPDIR.
+       create: boolean to determine if we should create folder (True)
+    '''
+    from sregistry.defaults import SREGISTRY_TMPDIR
+
+    # First priority for the base goes to the user requested.
+    tmpdir = requested_tmpdir or SREGISTRY_TMPDIR
+
+    prefix = prefix or "sregistry-tmp"
+    prefix = "%s.%s" %(prefix, next(tempfile._get_candidate_names()))
+    tmpdir = os.path.join(tmpdir, prefix)
+
+    if not os.path.exists(tmpdir) and create is True:
+        os.mkdir(tmpdir)
+
+    return tmpdir
+
 def get_userhome():
     '''get the user home based on the effective uid
     '''
     return pwd.getpwuid(os.getuid())[5]
 
 
-############################################################################
-## COMPRESSION #############################################################
-############################################################################
+################################################################################
+## COMPRESSION #################################################################
+################################################################################
 
 def extract_tar(archive, output_folder, handle_whiteout=False):
     '''extract a tar archive to a specified output folder
 
-    Parameters
-    ==========
-    archive: the archive file to extract
-    output_folder: the output folder to extract to
-    handle_whiteout: use docker2oci variation to handle whiteout files
+        Parameters
+        ==========
+        archive: the archive file to extract
+        output_folder: the output folder to extract to
+        handle_whiteout: use docker2oci variation to handle whiteout files
 
     '''
     from .terminal import run_command
@@ -122,10 +174,10 @@ def _extract_tar(archive, output_folder):
 
 def create_tar(files, output_folder=None):
     '''create_memory_tar will take a list of files (each a dictionary
-    with name, permission, and content) and write the tarfile
-    (a sha256 sum name is used) to the output_folder.
-    If there is no output folde specified, the
-    tar is written to a temporary folder.
+        with name, permission, and content) and write the tarfile
+        (a sha256 sum name is used) to the output_folder.
+        If there is no output folde specified, the
+        tar is written to a temporary folder.
     '''
     if output_folder is None:
         output_folder = tempfile.mkdtemp()
@@ -198,22 +250,28 @@ def get_content_hash(contents):
     return hasher.hexdigest()
 
 
-############################################################################
-## FILE OPERATIONS #########################################################
-############################################################################
+################################################################################
+## FILE OPERATIONS #############################################################
+################################################################################
 
 def copyfile(source, destination, force=True):
     '''copy a file from a source to its destination.
     '''
+    # Case 1: It's already there, we aren't replacing it :)
+    if source == destination and force is False:
+        return destination
+
+    # Case 2: It's already there, we ARE replacing it :)
     if os.path.exists(destination) and force is True:
         os.remove(destination)
+
     shutil.copyfile(source, destination)
     return destination
 
 
 def write_file(filename, content, mode="w"):
     '''write_file will open a file, "filename" and write content, "content"
-    and properly close the file
+       and properly close the file
     '''
     with open(filename, mode) as filey:
         filey.writelines(content)
@@ -222,9 +280,12 @@ def write_file(filename, content, mode="w"):
 
 def write_json(json_obj, filename, mode="w", print_pretty=True):
     '''write_json will (optionally,pretty print) a json object to file
-    :param json_obj: the dict to print to json
-    :param filename: the output file to write to
-    :param pretty_print: if True, will use nicer formatting
+
+       Parameters
+       ==========
+       json_obj: the dict to print to json
+       filename: the output file to write to
+       pretty_print: if True, will use nicer formatting
     '''
     with open(filename, mode) as filey:
         if print_pretty:
@@ -247,7 +308,7 @@ def print_json(json_obj):
 
 def read_file(filename, mode="r", readlines=True):
     '''write_file will open a file, "filename" and write content, "content"
-    and properly close the file
+       and properly close the file
     '''
     with open(filename, mode) as filey:
         if readlines is True:
@@ -259,7 +320,7 @@ def read_file(filename, mode="r", readlines=True):
 
 def read_json(filename, mode='r'):
     '''read_json reads in a json file and returns
-    the data structure as dict.
+       the data structure as dict.
     '''
     with open(filename, mode) as filey:
         data = json.load(filey)
